backend/requirements.txt
# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.1

# --- Database ---
motor==3.5.1          # Async MongoDB driver
pymongo==4.8.0        # Required for change streams

# --- Auth & Security ---
python-dotenv==1.0.1
requests==2.32.3

# --- AI Integration ---
google-generativeai==0.8.3

# --- Background Scheduler ---
apscheduler==3.10.4

# --- WebSockets ---
websockets==12.0
httpx==0.27.0

# --- Data Validation ---
pydantic==1.10.18     # FastAPI still depends on v1 for now
